+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [2.0.1] - 2018-12-12
+- Fixed "Fatal error: Uncaught Exception: MerchantID not valid" in test meta box.
+- Use issuer field from core gateway.
+- Updated deprecated function calls.
+
 ## [2.0.0] - 2018-05-09
 - Switched to PHP namespaces.
 
@@ -55,7 +60,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2015-01-19
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/icepay/compare/2.0.0...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/icepay/compare/2.0.1...HEAD
+[2.0.1]: https://github.com/wp-pay-gateways/icepay/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/wp-pay-gateways/icepay/compare/1.3.1...2.0.0
 [1.3.1]: https://github.com/wp-pay-gateways/icepay/compare/1.3.0...1.3.1
 [1.3.0]: https://github.com/wp-pay-gateways/icepay/compare/1.2.9...1.3.0